test: add ci pipline for testing
