+++ backend/.env.ci
@@ -1,0 +1,9 @@
+SECRET_KEY=django-insecure-ngyj-asc8h%&rge1y6ge#y3is)1t_b_8-3z28^rw(qji^#$rb0
+DEBUG=True
+ALLOWED_HOSTS=localhost, 127.0.0.1
+# Database settings
+POSTGRES_DB=peakperformance_db
+POSTGRES_USER=peakperformance
+POSTGRES_PASSWORD=chickenjockey
+DB_HOST=db
+DB_PORT=5432
